[proliferate] move plans and reviews domains
