+++ anyharness/crates/anyharness-lib/src/acp/manager.rs
@@ -13,8 +13,8 @@ use super::session_actor::{
 };
 use crate::api::http::latency::{latency_trace_fields, LatencyRequestContext};
 use crate::domains::agents::model::ResolvedAgent;
-use crate::plans::service::PlanService;
-use crate::reviews::service::ReviewService;
+use crate::domains::plans::service::PlanService;
+use crate::domains::reviews::service::ReviewService;
 use crate::sessions::attachment_storage::PromptAttachmentStorage;
 use crate::sessions::mcp::SessionMcpServer;
 use crate::sessions::model::SessionRecord;
@@ -428,8 +428,8 @@ mod tests {
         ResolvedArtifact,
     };
     use crate::domains::agents::registry::built_in_registry;
+    use crate::domains::plans::{service::PlanService, store::PlanStore};
     use crate::persistence::Db;
-    use crate::plans::{service::PlanService, store::PlanStore};
     use crate::sessions::model::{SessionEventRecord, SessionRecord};
     use crate::sessions::runtime_event::RuntimeInjectedSessionEvent;
     use crate::sessions::store::SessionStore;
