+++ anyharness/crates/anyharness-lib/src/app/mod.rs
@@ -13,19 +13,19 @@ use crate::cowork::store::CoworkStore;
 use crate::domains::agents::catalog::{LaunchCatalogService, ModelCatalogService};
 use crate::domains::agents::reconcile_execution::AgentReconcileService;
 use crate::domains::agents::seed::AgentSeedStore;
+use crate::domains::plans::runtime::PlanRuntime;
+use crate::domains::plans::service::PlanService;
+use crate::domains::plans::store::PlanStore;
+use crate::domains::reviews::hooks::ReviewSessionHooks;
+use crate::domains::reviews::mcp_auth::ReviewMcpAuth;
+use crate::domains::reviews::runtime::ReviewRuntime;
+use crate::domains::reviews::service::ReviewService;
+use crate::domains::reviews::store::ReviewStore;
 use crate::mobility::service::MobilityService;
 use crate::mobility::store::MobilityStore;
 use crate::persistence::Db;
-use crate::plans::runtime::PlanRuntime;
-use crate::plans::service::PlanService;
-use crate::plans::store::PlanStore;
 use crate::repo_roots::service::RepoRootService;
 use crate::repo_roots::store::RepoRootStore;
-use crate::reviews::hooks::ReviewSessionHooks;
-use crate::reviews::mcp_auth::ReviewMcpAuth;
-use crate::reviews::runtime::ReviewRuntime;
-use crate::reviews::service::ReviewService;
-use crate::reviews::store::ReviewStore;
 use crate::sessions::attachment_storage::PromptAttachmentStorage;
 use crate::sessions::links::completions::LinkCompletionStore;
 use crate::sessions::links::service::SessionLinkService;
