+++ anyharness/crates/anyharness-lib/src/domains/mod.rs
@@ -1,1 +1,3 @@
 pub mod agents;
+pub mod plans;
+pub mod reviews;
