+++ anyharness/crates/anyharness-lib/src/domains/reviews/service.rs
@@ -1,0 +1,655 @@
+use anyharness_contract::v1;
+use uuid::Uuid;
+
+use super::model::{
+    ReviewAssignmentStatus, ReviewCodeTargetManifest, ReviewFeedbackJobRecord,
+    ReviewFeedbackJobState, ReviewKind, ReviewModeVerificationStatus, ReviewRoundRecord,
+    ReviewRoundStatus, ReviewRunRecord, ReviewRunStatus,
+};
+use super::service_detail::{
+    assignment_is_terminal, build_assignments, build_feedback_prompt, dedupe_personas,
+    map_link_error, session_has_review_mcp, validate_review_submission, validate_reviewers,
+    validate_rounds,
+};
+use super::store::ReviewStore;
+use crate::domains::plans::model::PlanRecord;
+use crate::domains::plans::service::PlanService;
+use crate::sessions::links::model::{SessionLinkRelation, SessionLinkWorkspaceRelation};
+use crate::sessions::links::service::{CreateSessionLinkInput, SessionLinkService};
+use crate::sessions::store::SessionStore;
+
+pub const REVIEWER_DEADLINE_MINUTES: i64 = 30;
+pub const MAX_REVIEWERS_PER_RUN: usize = 4;
+pub const MAX_REVIEW_ROUNDS: u32 = 10;
+pub const MAX_REVIEW_SUMMARY_BYTES: usize = 4 * 1024;
+pub const MAX_REVIEW_CRITIQUE_BYTES: usize = 64 * 1024;
+
+#[derive(Debug, Clone)]
+pub struct ReviewPersonaInput {
+    pub persona_id: String,
+    pub label: String,
+    pub prompt: String,
+    pub agent_kind: String,
+    pub model_id: Option<String>,
+    pub mode_id: Option<String>,
+}
+
+#[derive(Debug, Clone)]
+pub struct StartReviewInput {
+    pub workspace_id: String,
+    pub parent_session_id: String,
+    pub kind: ReviewKind,
+    pub title: String,
+    pub target_plan: Option<PlanRecord>,
+    pub target_code_manifest: Option<ReviewCodeTargetManifest>,
+    pub max_rounds: u32,
+    pub auto_iterate: bool,
+    pub reviewers: Vec<ReviewPersonaInput>,
+}
+
+#[derive(Debug, thiserror::Error)]
+pub enum ReviewError {
+    #[error("session not found: {0}")]
+    SessionNotFound(String),
+    #[error("workspace not found: {0}")]
+    WorkspaceNotFound(String),
+    #[error("plan not found: {0}")]
+    PlanNotFound(String),
+    #[error("plan does not belong to this review parent session")]
+    PlanParentMismatch,
+    #[error("review run not found: {0}")]
+    RunNotFound(String),
+    #[error("review assignment not found: {0}")]
+    AssignmentNotFound(String),
+    #[error("a review is already active for this session")]
+    ActiveReviewExists,
+    #[error("reviewer list must contain between 1 and {MAX_REVIEWERS_PER_RUN} reviewers")]
+    InvalidReviewerCount,
+    #[error("max rounds must be between 1 and {MAX_REVIEW_ROUNDS}")]
+    InvalidMaxRounds,
+    #[error("review run is not waiting for a revision")]
+    NotWaitingForRevision,
+    #[error("review run has reached its max rounds")]
+    MaxRoundsReached,
+    #[error("revised plan is required")]
+    RevisedPlanRequired,
+    #[error("multiple revised plans are available; pass revisedPlanId")]
+    AmbiguousRevisedPlan,
+    #[error("cannot submit review result after assignment is terminal")]
+    AssignmentTerminal,
+    #[error("review assignment cannot be retried in its current state")]
+    RetryNotAllowed,
+    #[error("review {0} is too large")]
+    ReviewSubmissionTooLarge(&'static str),
+    #[error("session link failed: {0}")]
+    Link(String),
+    #[error(transparent)]
+    Internal(anyhow::Error),
+}
+
+#[derive(Clone)]
+pub struct ReviewService {
+    store: ReviewStore,
+    session_store: SessionStore,
+    link_service: SessionLinkService,
+    plan_service: ArcPlanService,
+}
+
+type ArcPlanService = std::sync::Arc<PlanService>;
+
+impl ReviewService {
+    pub fn new(
+        store: ReviewStore,
+        session_store: SessionStore,
+        link_service: SessionLinkService,
+        plan_service: ArcPlanService,
+    ) -> Self {
+        Self {
+            store,
+            session_store,
+            link_service,
+            plan_service,
+        }
+    }
+
+    pub fn store(&self) -> &ReviewStore {
+        &self.store
+    }
+
+    pub fn get_plan(&self, plan_id: &str) -> anyhow::Result<Option<PlanRecord>> {
+        self.plan_service.get(plan_id)
+    }
+
+    pub fn run_accepts_manual_revision_signal(&self, run: &ReviewRunRecord) -> bool {
+        if run.current_round_number >= run.max_rounds {
+            return false;
+        }
+        matches!(run.status, ReviewRunStatus::WaitingForRevision)
+            || (run.status == ReviewRunStatus::ParentRevising && !run.auto_iterate)
+    }
+
+    pub fn run_can_signal_revision_via_mcp(&self, run: &ReviewRunRecord) -> bool {
+        run.parent_can_signal_revision_via_mcp && self.run_accepts_manual_revision_signal(run)
+    }
+
+    pub fn start_review(&self, input: StartReviewInput) -> Result<ReviewRunRecord, ReviewError> {
+        validate_rounds(input.max_rounds)?;
+        validate_reviewers(&input.reviewers)?;
+        let parent = self
+            .session_store
+            .find_by_id(&input.parent_session_id)
+            .map_err(ReviewError::Internal)?
+            .ok_or_else(|| ReviewError::SessionNotFound(input.parent_session_id.clone()))?;
+        if parent.workspace_id != input.workspace_id {
+            return Err(ReviewError::WorkspaceNotFound(input.workspace_id));
+        }
+        if self
+            .store
+            .find_active_run_for_parent(&input.parent_session_id)
+            .map_err(ReviewError::Internal)?
+            .is_some()
+        {
+            return Err(ReviewError::ActiveReviewExists);
+        }
+        if let Some(plan) = input.target_plan.as_ref() {
+            if plan.workspace_id != input.workspace_id || plan.session_id != input.parent_session_id
+            {
+                return Err(ReviewError::PlanParentMismatch);
+            }
+        }
+
+        let now = chrono::Utc::now().to_rfc3339();
+        let run_id = Uuid::new_v4().to_string();
+        let round_id = Uuid::new_v4().to_string();
+        let target_code_manifest_json = input
+            .target_code_manifest
+            .as_ref()
+            .map(serde_json::to_string)
+            .transpose()
+            .map_err(|error| ReviewError::Internal(anyhow::Error::from(error)))?;
+        let parent_can_signal_revision_via_mcp = session_has_review_mcp(&parent);
+        let run = ReviewRunRecord {
+            id: run_id.clone(),
+            workspace_id: input.workspace_id.clone(),
+            parent_session_id: input.parent_session_id.clone(),
+            kind: input.kind,
+            status: ReviewRunStatus::Reviewing,
+            target_plan_id: input.target_plan.as_ref().map(|plan| plan.id.clone()),
+            target_plan_snapshot_hash: input
+                .target_plan
+                .as_ref()
+                .map(|plan| plan.snapshot_hash.clone()),
+            target_code_manifest_json: target_code_manifest_json.clone(),
+            title: input.title,
+            max_rounds: input.max_rounds,
+            auto_iterate: input.auto_iterate,
+            active_round_id: Some(round_id.clone()),
+            current_round_number: 1,
+            parent_can_signal_revision_via_mcp,
+            failure_reason: None,
+            failure_detail: None,
+            stopped_at: None,
+            created_at: now.clone(),
+            updated_at: now.clone(),
+        };
+        let round = ReviewRoundRecord {
+            id: round_id.clone(),
+            review_run_id: run_id.clone(),
+            round_number: 1,
+            status: ReviewRoundStatus::Reviewing,
+            target_plan_id: run.target_plan_id.clone(),
+            target_plan_snapshot_hash: run.target_plan_snapshot_hash.clone(),
+            target_code_manifest_json,
+            feedback_job_id: None,
+            feedback_prompt_sent_at: None,
+            completed_at: None,
+            failure_reason: None,
+            failure_detail: None,
+            created_at: now.clone(),
+            updated_at: now.clone(),
+        };
+        let assignments = build_assignments(&run, &round, &input.reviewers, &now);
+        self.store
+            .create_run(&run, &round, &assignments)
+            .map_err(ReviewError::Internal)?;
+        Ok(run)
+    }
+
+    pub fn link_reviewer_session(
+        &self,
+        run_id: &str,
+        assignment_id: &str,
+        parent_session_id: &str,
+        reviewer_session_id: &str,
+        label: Option<String>,
+        actual_mode_id: Option<&str>,
+        mode_status: ReviewModeVerificationStatus,
+    ) -> Result<String, ReviewError> {
+        let link = self
+            .link_service
+            .create_link(CreateSessionLinkInput {
+                relation: SessionLinkRelation::ReviewAgent,
+                parent_session_id: parent_session_id.to_string(),
+                child_session_id: reviewer_session_id.to_string(),
+                workspace_relation: SessionLinkWorkspaceRelation::SameWorkspace,
+                label,
+                created_by_turn_id: None,
+                created_by_tool_call_id: None,
+            })
+            .map_err(map_link_error)?;
+        let launched = self
+            .store
+            .update_assignment_launched(
+                assignment_id,
+                reviewer_session_id,
+                &link.id,
+                actual_mode_id,
+                mode_status,
+            )
+            .map_err(ReviewError::Internal)?;
+        if !launched {
+            self.link_service
+                .delete_link(&link.id)
+                .map_err(ReviewError::Internal)?;
+            return Err(ReviewError::RetryNotAllowed);
+        }
+        tracing::info!(
+            review_run_id = %run_id,
+            assignment_id,
+            reviewer_session_id,
+            session_link_id = %link.id,
+            "linked review agent session"
+        );
+        Ok(link.id)
+    }
+
+    pub fn submit_assignment_result(
+        &self,
+        reviewer_session_id: &str,
+        pass: bool,
+        summary: &str,
+        critique_markdown: &str,
+        critique_artifact_path: &str,
+    ) -> Result<Option<ReviewFeedbackJobRecord>, ReviewError> {
+        validate_review_submission(summary, critique_markdown)?;
+        let assignment = self
+            .store
+            .find_assignment_for_reviewer_session(reviewer_session_id)
+            .map_err(ReviewError::Internal)?
+            .ok_or_else(|| ReviewError::AssignmentNotFound(reviewer_session_id.to_string()))?;
+        let Some(updated) = self
+            .store
+            .submit_assignment_result(
+                &assignment.id,
+                pass,
+                summary,
+                critique_markdown,
+                critique_artifact_path,
+            )
+            .map_err(ReviewError::Internal)?
+        else {
+            return Err(ReviewError::AssignmentTerminal);
+        };
+        self.try_complete_round(&updated.review_round_id)
+    }
+
+    pub fn try_complete_round(
+        &self,
+        round_id: &str,
+    ) -> Result<Option<ReviewFeedbackJobRecord>, ReviewError> {
+        let assignments = self
+            .store
+            .list_assignments_for_round(round_id)
+            .map_err(ReviewError::Internal)?;
+        if assignments.is_empty() {
+            return Ok(None);
+        }
+        if assignments
+            .iter()
+            .any(|assignment| !assignment_is_terminal(assignment.status))
+        {
+            return Ok(None);
+        }
+        if !self
+            .store
+            .claim_round_for_completion(round_id)
+            .map_err(ReviewError::Internal)?
+        {
+            return Ok(None);
+        }
+        let round = self
+            .store
+            .find_round(round_id)
+            .map_err(ReviewError::Internal)?
+            .ok_or_else(|| ReviewError::RunNotFound(round_id.to_string()))?;
+        let run = self
+            .store
+            .find_run(&round.review_run_id)
+            .map_err(ReviewError::Internal)?
+            .ok_or_else(|| ReviewError::RunNotFound(round.review_run_id.clone()))?;
+        let submitted_count = assignments
+            .iter()
+            .filter(|assignment| assignment.status == ReviewAssignmentStatus::Submitted)
+            .count();
+        if submitted_count == 0 {
+            self.store
+                .mark_run_system_failed(
+                    &run.id,
+                    Some(round_id),
+                    "all_reviewers_failed",
+                    Some("No reviewer submitted critique before the review round ended."),
+                )
+                .map_err(ReviewError::Internal)?;
+            return Ok(None);
+        }
+
+        let now = chrono::Utc::now().to_rfc3339();
+        let all_approved = assignments.iter().all(|assignment| {
+            assignment.status == ReviewAssignmentStatus::Submitted && assignment.pass == Some(true)
+        });
+        if all_approved {
+            let job = ReviewFeedbackJobRecord {
+                id: Uuid::new_v4().to_string(),
+                review_run_id: run.id.clone(),
+                review_round_id: round.id.clone(),
+                parent_session_id: run.parent_session_id.clone(),
+                state: ReviewFeedbackJobState::Pending,
+                prompt_text: build_feedback_prompt(&run, &round, &assignments),
+                attempt_count: 0,
+                next_attempt_at: None,
+                sent_prompt_seq: None,
+                feedback_turn_id: None,
+                failure_reason: None,
+                failure_detail: None,
+                created_at: now.clone(),
+                updated_at: now,
+            };
+            self.store
+                .create_feedback_job(&job, ReviewRunStatus::Passed)
+                .map_err(ReviewError::Internal)?;
+            return Ok(Some(job));
+        }
+
+        let job = ReviewFeedbackJobRecord {
+            id: Uuid::new_v4().to_string(),
+            review_run_id: run.id.clone(),
+            review_round_id: round.id.clone(),
+            parent_session_id: run.parent_session_id.clone(),
+            state: ReviewFeedbackJobState::Pending,
+            prompt_text: build_feedback_prompt(&run, &round, &assignments),
+            attempt_count: 0,
+            next_attempt_at: None,
+            sent_prompt_seq: None,
+            feedback_turn_id: None,
+            failure_reason: None,
+            failure_detail: None,
+            created_at: now.clone(),
+            updated_at: now,
+        };
+        self.store
+            .create_feedback_job(&job, ReviewRunStatus::FeedbackReady)
+            .map_err(ReviewError::Internal)?;
+        Ok(Some(job))
+    }
+
+    pub fn list_session_reviews(
+        &self,
+        session_id: &str,
+    ) -> Result<Vec<v1::ReviewRunDetail>, ReviewError> {
+        self.session_store
+            .find_by_id(session_id)
+            .map_err(ReviewError::Internal)?
+            .ok_or_else(|| ReviewError::SessionNotFound(session_id.to_string()))?;
+        let runs = self
+            .store
+            .list_runs_for_parent(session_id)
+            .map_err(ReviewError::Internal)?;
+        runs.iter().map(|run| self.detail_for_run(run)).collect()
+    }
+
+    pub fn get_run_detail(&self, run_id: &str) -> Result<v1::ReviewRunDetail, ReviewError> {
+        let run = self
+            .store
+            .find_run(run_id)
+            .map_err(ReviewError::Internal)?
+            .ok_or_else(|| ReviewError::RunNotFound(run_id.to_string()))?;
+        self.detail_for_run(&run)
+    }
+
+    pub fn get_assignment_critique(
+        &self,
+        run_id: &str,
+        assignment_id: &str,
+    ) -> Result<v1::ReviewCritiqueResponse, ReviewError> {
+        let assignment = self
+            .store
+            .find_assignment_for_run(run_id, assignment_id)
+            .map_err(ReviewError::Internal)?
+            .ok_or_else(|| ReviewError::AssignmentNotFound(assignment_id.to_string()))?;
+        Ok(v1::ReviewCritiqueResponse {
+            assignment_id: assignment.id,
+            review_run_id: assignment.review_run_id,
+            review_round_id: assignment.review_round_id,
+            persona_id: assignment.persona_id,
+            persona_label: assignment.persona_label,
+            pass: assignment.pass,
+            summary: assignment.summary,
+            critique_markdown: assignment.critique_markdown,
+            critique_artifact_path: assignment.critique_artifact_path,
+            submitted_at: assignment.submitted_at,
+        })
+    }
+
+    pub fn stop_run(&self, run_id: &str) -> Result<Vec<String>, ReviewError> {
+        self.store.stop_run(run_id).map_err(ReviewError::Internal)
+    }
+
+    pub(crate) fn delete_unlaunched_reviewer_session(
+        &self,
+        session_id: &str,
+    ) -> Result<(), ReviewError> {
+        self.session_store
+            .delete_session(session_id)
+            .map_err(ReviewError::Internal)
+    }
+
+    pub fn mark_parent_feedback_turn_finished(
+        &self,
+        parent_session_id: &str,
+        turn_id: &str,
+    ) -> Result<Option<ReviewRunRecord>, ReviewError> {
+        let run_id = self
+            .store
+            .mark_parent_feedback_turn_finished(parent_session_id, turn_id)
+            .map_err(ReviewError::Internal)?;
+        run_id
+            .map(|run_id| {
+                self.store
+                    .find_run(&run_id)
+                    .map_err(ReviewError::Internal)?
+                    .ok_or_else(|| ReviewError::RunNotFound(run_id))
+            })
+            .transpose()
+    }
+
+    pub fn record_candidate_plan(&self, plan: &PlanRecord) {
+        let Ok(Some(run)) = self.store.find_active_run_for_parent(&plan.session_id) else {
+            return;
+        };
+        if !matches!(
+            run.status,
+            ReviewRunStatus::ParentRevising | ReviewRunStatus::WaitingForRevision
+        ) {
+            return;
+        }
+        if let Err(error) = self.store.record_candidate_plan(
+            &run.id,
+            &plan.id,
+            plan.source_turn_id.as_deref(),
+            plan.source_tool_call_id.as_deref(),
+            &plan.snapshot_hash,
+        ) {
+            tracing::warn!(review_run_id = %run.id, plan_id = %plan.id, error = %error, "failed to record review candidate plan");
+        }
+    }
+
+    pub fn start_next_round_records(
+        &self,
+        run_id: &str,
+        revised_plan_id: Option<&str>,
+        target_manifest: Option<ReviewCodeTargetManifest>,
+    ) -> Result<ReviewRunRecord, ReviewError> {
+        self.start_next_round_records_with_claim(run_id, revised_plan_id, target_manifest, None)
+            .map(|(run, _started)| run)
+    }
+
+    pub fn start_next_round_records_after_feedback_turn(
+        &self,
+        run_id: &str,
+        feedback_turn_id: &str,
+        revised_plan_id: Option<&str>,
+        target_manifest: Option<ReviewCodeTargetManifest>,
+    ) -> Result<(ReviewRunRecord, bool), ReviewError> {
+        self.start_next_round_records_with_claim(
+            run_id,
+            revised_plan_id,
+            target_manifest,
+            Some(feedback_turn_id),
+        )
+    }
+
+    fn start_next_round_records_with_claim(
+        &self,
+        run_id: &str,
+        revised_plan_id: Option<&str>,
+        target_manifest: Option<ReviewCodeTargetManifest>,
+        feedback_turn_id: Option<&str>,
+    ) -> Result<(ReviewRunRecord, bool), ReviewError> {
+        let run = self
+            .store
+            .find_run(run_id)
+            .map_err(ReviewError::Internal)?
+            .ok_or_else(|| ReviewError::RunNotFound(run_id.to_string()))?;
+        if !matches!(
+            run.status,
+            ReviewRunStatus::FeedbackReady
+                | ReviewRunStatus::ParentRevising
+                | ReviewRunStatus::WaitingForRevision
+        ) {
+            if feedback_turn_id.is_some() {
+                return Ok((run, false));
+            }
+            return Err(ReviewError::NotWaitingForRevision);
+        }
+        if run.current_round_number >= run.max_rounds {
+            if matches!(
+                run.status,
+                ReviewRunStatus::ParentRevising | ReviewRunStatus::WaitingForRevision
+            ) {
+                self.store
+                    .mark_run_max_rounds_reached(&run.id)
+                    .map_err(ReviewError::Internal)?;
+                return self
+                    .store
+                    .find_run(&run.id)
+                    .map_err(ReviewError::Internal)?
+                    .ok_or_else(|| ReviewError::RunNotFound(run.id))
+                    .map(|run| (run, false));
+            }
+            return Err(ReviewError::MaxRoundsReached);
+        }
+
+        let revised_plan = match run.kind {
+            ReviewKind::Plan => {
+                let inferred_plan_id;
+                let plan_id = match revised_plan_id {
+                    Some(plan_id) => plan_id,
+                    None => {
+                        inferred_plan_id = self
+                            .store
+                            .find_single_candidate_plan_id(&run.id, feedback_turn_id)
+                            .map_err(|error| {
+                                if error.to_string()
+                                    == "multiple revised plan candidates are available"
+                                {
+                                    ReviewError::AmbiguousRevisedPlan
+                                } else {
+                                    ReviewError::Internal(error)
+                                }
+                            })?
+                            .ok_or(ReviewError::RevisedPlanRequired)?;
+                        inferred_plan_id.as_str()
+                    }
+                };
+                let plan = self
+                    .plan_service
+                    .get(plan_id)
+                    .map_err(ReviewError::Internal)?
+                    .ok_or_else(|| ReviewError::PlanNotFound(plan_id.to_string()))?;
+                if plan.workspace_id != run.workspace_id || plan.session_id != run.parent_session_id
+                {
+                    return Err(ReviewError::PlanParentMismatch);
+                }
+                Some(plan)
+            }
+            ReviewKind::Code => None,
+        };
+        let previous_assignments = self
+            .store
+            .list_assignments_for_run(run_id)
+            .map_err(ReviewError::Internal)?;
+        let personas = dedupe_personas(previous_assignments);
+        let now = chrono::Utc::now().to_rfc3339();
+        let round_number = run.current_round_number + 1;
+        let round = ReviewRoundRecord {
+            id: Uuid::new_v4().to_string(),
+            review_run_id: run.id.clone(),
+            round_number,
+            status: ReviewRoundStatus::Reviewing,
+            target_plan_id: revised_plan.as_ref().map(|plan| plan.id.clone()),
+            target_plan_snapshot_hash: revised_plan.as_ref().map(|plan| plan.snapshot_hash.clone()),
+            target_code_manifest_json: target_manifest
+                .as_ref()
+                .map(serde_json::to_string)
+                .transpose()
+                .map_err(|error| ReviewError::Internal(anyhow::Error::from(error)))?,
+            feedback_job_id: None,
+            feedback_prompt_sent_at: None,
+            completed_at: None,
+            failure_reason: None,
+            failure_detail: None,
+            created_at: now.clone(),
+            updated_at: now.clone(),
+        };
+        let assignments = build_assignments(&run, &round, &personas, &now);
+        let started = self
+            .store
+            .start_next_round(
+                &run.id,
+                &round,
+                &assignments,
+                round.target_plan_id.as_deref(),
+                round.target_plan_snapshot_hash.as_deref(),
+                round.target_code_manifest_json.as_deref(),
+                run.current_round_number,
+                feedback_turn_id,
+            )
+            .map_err(ReviewError::Internal)?;
+        if !started {
+            return self
+                .store
+                .find_run(&run.id)
+                .map_err(ReviewError::Internal)?
+                .ok_or_else(|| ReviewError::RunNotFound(run.id))
+                .map(|run| (run, false));
+        }
+        self.store
+            .find_run(&run.id)
+            .map_err(ReviewError::Internal)?
+            .ok_or_else(|| ReviewError::RunNotFound(run.id))
+            .map(|run| (run, true))
+    }
+}
+
+#[cfg(test)]
+mod service_tests;
