+++ anyharness/crates/anyharness-lib/src/lib.rs
@@ -8,9 +8,7 @@ pub mod integrations;
 pub mod mobility;
 pub mod origin;
 pub mod persistence;
-pub mod plans;
 pub mod repo_roots;
-pub mod reviews;
 pub mod sessions;
 pub mod terminals;
 pub mod workspaces;
